Add UI volume setting (affecting navigation sounds and preview audio). Convert alt speed volume into a 100-step percentage.

+++ Common/System/System.h
@@ -302,8 +302,8 @@ void System_AudioClear();
 // These samples really have 16 bits of value, but can be a little out of range.
 // This is for pushing rate-controlled 44khz audio from emulation.
 // If you push a little too fast, we'll pitch up to a limit, for example.
-// Volume is a 12-bit multiplier.
-void System_AudioPushSamples(const int32_t *audio, int numSamples, int volume);
+// Volume is a unit-range multiplier.
+void System_AudioPushSamples(const int32_t *audio, int numSamples, float volume);
 
 inline void System_AudioResetStatCounters() {
 	return System_AudioGetDebugStats(nullptr, 0);
