+++ UI/AudioCommon.cpp
@@ -28,7 +28,7 @@ void System_AudioClear() {
 	g_resampler.Clear();
 }
 
-void System_AudioPushSamples(const int32_t *audio, int numSamples, int volume) {
+void System_AudioPushSamples(const int32_t *audio, int numSamples, float volume) {
 	if (audio) {
 		g_resampler.PushSamples(audio, numSamples, volume);
 	} else {
